fix heavy sync test
