+++ crates/p2p/src/tests/block_pool/mod.rs
@@ -1,7 +1,7 @@
 use crate::block_pool_service::{BlockPoolService, ProcessBlock};
 use crate::peer_list::{AddPeer, PeerListServiceWithClient};
 use crate::tests::util::{FakeGossipServer, MockRethServiceActor};
-use crate::{sync_chain, SyncState};
+use crate::{SyncState};
 use actix::Actor;
 use async_trait::async_trait;
 use base58::ToBase58;
@@ -131,6 +131,7 @@ async fn should_process_block() {
     let peer_addr = peer_list_service.start();
 
     let (vdf_tx, _vdf_rx) = tokio::sync::mpsc::channel(1);
+    let (vdf_service_tx, _vdf_service_rx) = tokio::sync::mpsc::unbounded_channel();
     let sync_state = SyncState::new(false);
     let service = BlockPoolService::new_with_client(
         db.clone(),
@@ -138,6 +139,7 @@ async fn should_process_block() {
         block_discovery_stub.clone(),
         Some(vdf_tx),
         sync_state,
+        vdf_service_tx
     );
     let addr = service.start();
 
@@ -287,6 +289,7 @@ async fn should_process_block_with_intermediate_block_in_api() {
         .expect("can't send message to peer list");
 
     let (vdf_tx, _vdf_rx) = tokio::sync::mpsc::channel(1);
+    let (vdf_service_tx, _vdf_service_rx) = tokio::sync::mpsc::unbounded_channel();
     let sync_state = SyncState::new(false);
 
     let service = BlockPoolService::new_with_client(
@@ -295,6 +298,7 @@ async fn should_process_block_with_intermediate_block_in_api() {
         block_discovery_stub.clone(),
         Some(vdf_tx),
         sync_state,
+        vdf_service_tx
     );
     let addr = service.start();
